refactor: 파일 구조 변경 및 dto이름 변경 (#5)
Refs: #5
