+++ BACK/spring-app/src/main/java/com/starchive/springapp/hashtag/dto/HashTagCheckRequest.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class HashTagRequest {
+public class HashTagCheckRequest {
     @Schema(description = "해시태그 이름", example = "Spring", maxLength = 32)
     @NotEmpty(message = "해쉬태그이름은 1글자 이상이어야 합니다.")
     @Size(max = 32, message = "해쉬태그이름은 32자 보다 작거나 같아야 합니다.")
