+++ BACK/spring-app/src/main/java/com/starchive/springapp/post/domain/Post.java
@@ -1,40 +1,54 @@
-package com.starchive.springapp.posthashtag;
+package com.starchive.springapp.post.domain;
 
 import static jakarta.persistence.FetchType.LAZY;
 
-import com.starchive.springapp.hashtag.domain.HashTag;
-import com.starchive.springapp.post.Post;
+import com.starchive.springapp.category.domain.Category;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "PostHashTag")
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 @AllArgsConstructor
-public class PostHashTag {
+@Table(name = "Posts")
+public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "postHashTagId")
+    @Column(name = "postId")
     private Long id;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "postId")
-    private Post post;
+    @Column(nullable = false, length = 64)
+    private String title;
+
+    @Lob
+    private String content;
+
+    @Column(nullable = false, length = 32)
+    private String author;
+
+    @Column(nullable = false, length = 128)
+    private String password;
+
+    @Column(nullable = false, name = "datetime")
+    private LocalDateTime dateTime;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "hashTagId")
-    private HashTag hashTag;
+    @JoinColumn(name = "categoryId")
+    private Category category;
 
 
 }
